Add files via upload
The Electronic Personal Aid is an AI-Powered Virtual Assistant. This retrieves answers and solutions from the free version of ChatGPT. There is also another customized AI-bot included in the project, which I created using the Botpress Cloud website. The Chrome driver helps us to have a connection with that website. I've used Python as both "Frontend" and "Backend". This virtual assistant is named "Vikcy". It doesn't exactly work like sophisticated assistants like "Google Assistant", "Siri", "Alexa" etc., but it does the job. I recommend to execute this project on the "PyCharm".